Rolling hard due dates tool plugin
This plugin gives a library the ability define a set of dates when
a hard due date will be updated. On that date, the hard due date
in the issuing rules will be set, and any issued items that are due
after that date will have their due dates reduced to the hard due date.